Move ResourceArray to its own header
 - Add docs.
